requirements.txt
fastapi
uvicorn
python-chess
torch
numpy
pydantic